Fix python error when printing a file that doesn't use the heated bed (#1)
...on a printer that does have a heated bed.

The error happens because `plugins.smartpreheat.bed` ends up being `None`, which can't be multiplied by a float. I'm not sure if this is new with Python 3.

I decided to skip the smart preheat bed commands altogether if no bed temperature is found in the file (rather than just adding `or 0` to the preheat commands), as this is what the printed file would have done without smart preheat.